This commit allows cutouts to continue if automatic redownloading fails.
Previously the cutout script would break if it could not successfully grab new images from NCSA. I allow it to continue since these missing images are expected to be few in number and can be replaced with NaNs. The effect will be shortened time series image sets, but the impact should be small.